requirements.txt
pydantic
litellm
rich
python-dotenv
anthropic
openai
groq